CRUD queries implemented;

+++ src/main/java/dao/CrudDao.java
@@ -6,9 +6,9 @@ import java.util.List;
 import java.util.Optional;
 
 public interface CrudDao<T extends IdHolder> {
-    Optional<T> find(int id);
+    Optional<T> find(long id);
     List<T> findAll();
     void create(T entity);
     void update(T entity);
-    void delete(int id);
+    void delete(long id);
 }
